requirements.txt
amqp==5.3.1
asttokens==3.0.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
checksumdir==1.2.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.2
debugpy==1.8.15
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
Flask==3.1.1
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kombu==5.5.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pyzmq==27.0.0
redis==5.2.1
requests==2.32.4
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
